requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
numpy==1.26.4
opencv-python==4.11.0.86
pillow==11.2.1
requests==2.32.4
torch==1.11.0+cpu
torchaudio==0.11.0+cpu
torchvision==0.12.0+cpu
typing_extensions==4.14.0
urllib3==2.5.0